+++ index.js
@@ -1,5 +1,0 @@
-'use strict';
-
-const { ElasticSearch } = require('./lib');
-
-module.exports = ElasticSearch;